@@ -1,17 +1,28 @@
-﻿using idCardXF.Views;
+﻿using BethanysPieShopStockApp.Services;
+using BethanysPieShopStockApp.Utility;
+using idCardXF.Repository;
+using idCardXF.Services;
+using idCardXF.Views;
 using System;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
+
+[assembly: ExportFont("Lobster-Regular.ttf", Alias = "Lobster")]
 
 namespace idCardXF
 {
     public partial class App : Application
     {
+        public static NavigationService NavigationService { get; } = new NavigationService();
+        public static IStudentService StudentService { get; set; } = new StudentService(new StudentRepository());
         public App()
         {
             InitializeComponent();
 
-            MainPage = new HomeView();
+            NavigationService.Configure(ViewNames.HomeView, typeof(HomeView));
+            NavigationService.Configure(ViewNames.CoursesView, typeof(CoursesView));
+
+            MainPage = new NavigationPage(new HomeView());
         }
 
         protected override void OnStart()
